+++ app/move/app/view/middleware/locals.move-details.js
@@ -1,9 +1,9 @@
 const presenters = require('../../../../../common/presenters')
 
 function localsMoveDetails(req, res, next) {
-  const { move } = req
+  const { journeys, move } = req
 
-  const moveDetails = presenters.moveToMetaListComponent(move)
+  const moveDetails = presenters.moveToMetaListComponent(move, journeys)
 
   res.locals.moveDetails = moveDetails
   res.locals.moveIsLockout = move.is_lockout