+++ common/services/whats-new-content.js
@@ -10,6 +10,8 @@ const {
 } = require('../../config')
 const { DATE_FORMATS } = require('../../config/index')
 
+const TWO_WEEKS = 2 * 7 * 24 * 60 * 60 * 1000
+
 const options = {
   renderMark: {
     [MARKS.BOLD]: text => `<strong>${text}</strong>`,
@@ -27,15 +29,18 @@ const options = {
         node.content
       )}</h3>`,
     [BLOCKS.HEADING_4]: (node, next) =>
-      `<h4 class="govuk-heading-s">${next(node.content)}</h4>`,
+      `<h4 class="govuk-heading-m">${next(node.content)}</h4>`,
     [BLOCKS.UL_LIST]: (node, next) =>
       `<ul class="govuk-list govuk-list--bullet">${next(node.content)}</ul>`,
+    [BLOCKS.OL_LIST]: (node, next) =>
+      `<ol class="govuk-list govuk-list--number">${next(node.content)}</ol>`,
     [BLOCKS.PARAGRAPH]: (node, next) =>
       `<p class="govuk-body">${next(node.content)}</p>`,
     [INLINES.HYPERLINK]: (node, next) =>
       `<a class="govuk-link" href="${node.data.uri}">${next(node.content)}</a>`,
   },
 }
+
 const service = {
   client: contentful.createClient({
     host: CONTENTFUL_HOST,
@@ -43,39 +48,49 @@ const service = {
     accessToken: CONTENTFUL_ACCESS_TOKEN,
   }),
   fetch: async () => {
-    const entries = await service.client.getEntries()
+    const entries = await service.fetchEntries()
 
-    if (!entries.items?.length) {
+    if (entries.length === 0) {
       return null
     }
 
-    const formattedEntries = entries.items.map(
-      ({ fields: { title, body, date } }) => ({
-        title,
-        body: service.convertToHTMLFormat(body),
-        date: service.formatDate(date),
-      })
-    )
+    const formattedEntries = entries.map(({ title, body, date }) => ({
+      title,
+      body: service.convertToHTMLFormat(body),
+      date: service.formatDate(date),
+    }))
 
-    const latestContent = entries.items[0]
-    const latestContentTitle = latestContent.fields.title
-    const latestContentBannerText = latestContent.fields.briefBannerText
+    let formattedBannerContent = null
+
+    if (new Date() - entries[0].date <= TWO_WEEKS) {
+      formattedBannerContent = {
+        body: entries[0].briefBannerText,
+        date: service.formatDate(entries[0].date),
+      }
+    }
 
     return {
-      bannerContent: {
-        title: latestContentTitle,
-        body: latestContentBannerText,
-        date: service.formatDate(latestContent.fields.date),
-      },
+      bannerContent: formattedBannerContent,
       posts: formattedEntries,
     }
   },
-  convertToHTMLFormat: contentBody => {
-    return documentToHtmlString(contentBody, options)
+  fetchEntries: async () => {
+    const entries = await service.client.getEntries()
+
+    if (!entries.items?.length) {
+      return []
+    }
+
+    return entries.items.map(
+      ({ fields: { title, body, briefBannerText, date } }) => ({
+        title,
+        body,
+        briefBannerText,
+        date: new Date(date),
+      })
+    )
   },
-  formatDate: date => {
-    const newDate = new Date(date)
-    return format(newDate, DATE_FORMATS.WITH_MONTH)
-  },
+  convertToHTMLFormat: body => documentToHtmlString(body, options),
+  formatDate: date => format(date, DATE_FORMATS.WITH_MONTH),
 }
 module.exports = service