+++ common/services/whats-new-content.test.js
@@ -1,6 +1,11 @@
+const { format } = require('date-fns')
 const sinon = require('sinon')
 
+const { DATE_FORMATS } = require('../../config')
+
 const whatsNewContentService = require('./whats-new-content')
+
+const todaysDate = new Date()
 
 const mockedResponse = {
   items: [
@@ -140,10 +145,52 @@ const mockedResponse = {
               ],
               data: {},
             },
+            {
+              nodeType: 'ordered-list',
+              content: [
+                {
+                  nodeType: 'list-item',
+                  content: [
+                    {
+                      nodeType: 'paragraph',
+                      content: [
+                        {
+                          nodeType: 'text',
+                          value: 'TEST LINE 1',
+                          marks: [],
+                          data: {},
+                        },
+                      ],
+                      data: {},
+                    },
+                  ],
+                  data: {},
+                },
+                {
+                  nodeType: 'list-item',
+                  content: [
+                    {
+                      nodeType: 'paragraph',
+                      content: [
+                        {
+                          nodeType: 'text',
+                          value: 'TEST LINE 2',
+                          marks: [],
+                          data: {},
+                        },
+                      ],
+                      data: {},
+                    },
+                  ],
+                  data: {},
+                },
+              ],
+              data: {},
+            },
           ],
         },
         briefBannerText: 'Some text briefly explaining the changes.',
-        date: '2022-03-04T00:00+00:00',
+        date: todaysDate.toISOString(),
       },
     },
   ],
@@ -162,12 +209,8 @@ describe('whatsNewContentService Service', function () {
     it('returns the formatted body', async function () {
       const formattedEntries = await whatsNewContentService.fetch()
       expect(formattedEntries.posts[0].body).to.equal(
-        '<h1 class="govuk-heading-l">The latest updates and improvements to Book a secure move.</h1><h2 class="govuk-heading-s govuk-!-margin-top-5">Test heading 2.</h2><h3 class="govuk-heading-m govuk-!-margin-top-4"><i>Test heading 3.</i></h3><h4 class="govuk-heading-s">Test heading 4.</h4><a class="govuk-link" href="https://google.com">Test Link</a><p class="govuk-body"><strong>Some random paragraph text.</strong></p><ul class="govuk-list govuk-list--bullet"><li><p class="govuk-body">TEST LINE 1</p></li><li><p class="govuk-body">TEST LINE 2</p></li></ul>'
+        '<h1 class="govuk-heading-l">The latest updates and improvements to Book a secure move.</h1><h2 class="govuk-heading-s govuk-!-margin-top-5">Test heading 2.</h2><h3 class="govuk-heading-m govuk-!-margin-top-4"><i>Test heading 3.</i></h3><h4 class="govuk-heading-m">Test heading 4.</h4><a class="govuk-link" href="https://google.com">Test Link</a><p class="govuk-body"><strong>Some random paragraph text.</strong></p><ul class="govuk-list govuk-list--bullet"><li><p class="govuk-body">TEST LINE 1</p></li><li><p class="govuk-body">TEST LINE 2</p></li></ul><ol class="govuk-list govuk-list--number"><li><p class="govuk-body">TEST LINE 1</p></li><li><p class="govuk-body">TEST LINE 2</p></li></ol>'
       )
-    })
-    it('calls the whatsNewContentService.createClient method', async function () {
-      const results = await whatsNewContentService.fetch()
-      expect(results.bannerContent.title).to.equal('Whats new today!')
     })
     it('returns the content title', async function () {
       const formattedEntries = await whatsNewContentService.fetch()
@@ -181,7 +224,9 @@ describe('whatsNewContentService Service', function () {
     })
     it('returns the formatted date', async function () {
       const formattedEntries = await whatsNewContentService.fetch()
-      expect(formattedEntries.posts[0].date).to.equal('4 March 2022')
+      expect(formattedEntries.posts[0].date).to.equal(
+        format(todaysDate, DATE_FORMATS.WITH_MONTH)
+      )
     })
   })