+++ test/e2e/population/edit.test.js
@@ -1,6 +1,6 @@
-import { pmuUserWithRegions } from './_roles'
-import { populationDashboard } from './_routes'
-import { page, populationDashboardPage, populationEditPage } from './pages'
+import { pmuUserWithRegions } from '../_roles'
+import { populationDashboard } from '../_routes'
+import { page, populationDashboardPage, populationEditPage } from '../pages'
 
 fixture('Population Editing').page(populationDashboard)
 test.before(async t => {